% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{clean_common}
\alias{clean_common}
\title{Cleaning operation applied to results and amplification csv files}
\usage{
clean_common(x)
}
\arguments{
\item{x}{data frame to be cleaned}
}
\description{
Cleaning operation applied to results and amplification csv files
}
