% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kraken.R
\name{kraken_add_taxonomy}
\alias{kraken_add_taxonomy}
\title{Add full taxonomy path to a Kraken report data frame}
\usage{
kraken_add_taxonomy(x)
}
\arguments{
\item{x}{A data frame containing the fields "rank_code", "scientific_name",
and "rank_level" with rows in the original order of the Kraken sample
report(s), as returned by \code{read_kraken_reports(files, complete = FALSE, ...)}.}
}
\description{
This function requires that rows are in the order of the original Kraken
report. Thus, it cannot be used on data frames returned by \code{read_kraken()}
using the \code{complete = TRUE} option.
}
\details{
You can apply this function to a data frame of Kraken reports for multiple
samples; however, doing so is typically inefficient due to needing to
recompute paths for taxa that appear in multiple samples. Instead, consider
generating taxonomy paths from the original database and joining these to
the kraken report data frame.
}
