% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kraken.R
\name{kraken_complete}
\alias{kraken_complete}
\title{Complete a set of Kraken reports by filling in zero counts (Not tested!)}
\usage{
kraken_complete(x, explicit = TRUE)
}
\arguments{
\item{x}{A data frame of Kraken sample reports}

\item{explicit}{Boolean passed to \code{tidyr::complete()}}
}
\description{
This function uses \code{tidyr::complete()} to add rows with zeros for taxa that
missing in some samples. Note: The resulting data frame will no longer be
compatible with \code{kraken_add_taxonomy()}.
}
