% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kraken.R
\name{kraken_remove_nonstandard_ranks}
\alias{kraken_remove_nonstandard_ranks}
\title{Remove nonstandard ranks from Kraken reports}
\usage{
kraken_remove_nonstandard_ranks(x)
}
\arguments{
\item{x}{A data frame of Kraken sample reports}
}
\description{
Remove nonstandard ranks from Kraken reports
}
