% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naoutils-package.R
\docType{package}
\name{naoutils-package}
\alias{naoutils}
\alias{naoutils-package}
\title{naoutils: Helper Functions for the NAO Team}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael McLaren \email{mike@securebio.org} (\href{https://orcid.org/0000-0003-1575-473X}{ORCID})

}
\keyword{internal}
