% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naowetlab-package.R
\docType{package}
\name{naowetlab-package}
\alias{naowetlab}
\alias{naowetlab-package}
\title{naowetlab: Helper Functions for the NAO Wet Lab Team}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael McLaren \email{mike@securebio.org} (\href{https://orcid.org/0000-0003-1575-473X}{ORCID})

}
\keyword{internal}
