% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kraken.R
\name{read_kraken_reports}
\alias{read_kraken_reports}
\title{Read one or more Kraken2 sample reports}
\usage{
read_kraken_reports(files, minimizers, id = "file", num_threads = 1)
}
\arguments{
\item{files}{Paths to files}

\item{minimizers}{Boolean specifying whether the files include minimizer
data}

\item{id}{A string or NULL. If a string, create column with that name for
storing the files or the names of the \code{files} vector}

\item{num_threads}{Parameter passed onto \code{readr::read_tsv()}}
}
\description{
This function can read one or more Kraken(2) sample reports, as well as
output of \code{kraken2-inspect}. Kraken2 reports with and without minimizer data
are supported.
}
\details{
The column names in the resulting data frame are "clade_fragments_percent",
"clade_fragments", "node_fragments", "minimizers_total",
"minimizers_distinct", "rank_code", "taxid", "scientific_name"; the
minimizer columns are only included if \code{minimizers = TRUE}.

This function should work on Kraken and Kraken2 reports as well as on the
output of \code{kraken2-inspect}, but we've not tested it on original Kraken
reports. Currently there is no check as to whether the minimizer data is in
the reports or not; if \code{minimizers} is set incorrectly, the function will
return without error but the reports will be parsed incorrectly.
\itemize{
\item \href{https://github.com/DerrickWood/kraken2/wiki/Manual#sample-report-output-format}{Kraken2 docs}
\item \href{https://ccb.jhu.edu/software/kraken/MANUAL.html#output-format}{Kraken docs}
}
}
