% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{read_qpcr_amplification_csv}
\alias{read_qpcr_amplification_csv}
\title{Read qPCR 'Amplification Data' table from a CSV file}
\usage{
read_qpcr_amplification_csv(file)
}
\arguments{
\item{file}{path to file; see \code{readr::read_csv()}}
}
\description{
Read qPCR 'Amplification Data' table from a CSV file
}
